created a user_id session
all the classes and flashcard sets and flashcards are unique to each user so that each user can only add stuff to their own account

+++ flashcard_delete.php
@@ -5,6 +5,7 @@ session_start();
 if ($_SERVER['REQUEST_METHOD'] == "GET")
   {
     require('flashme-connectdb.php');
+    $user_id = $_SESSION['session_user_id'];
     $class_id = $_GET['class_id'];
     $set_id = $_GET['set_id'];
     $card_id = $_GET['card_id'];
@@ -14,7 +15,7 @@ if ($_SERVER['REQUEST_METHOD'] == "GET")
     // echo "There is already an account with that email";
     // }
 
-    $query = "DELETE FROM `flashcard` WHERE class_id=$class_id AND set_id=$set_id AND card_id=$card_id";
+    $query = "DELETE FROM `flashcard` WHERE user_id=$user_id AND class_id=$class_id AND set_id=$set_id AND card_id=$card_id";
     $result = $db->query($query);
 
     //echo "set_id=$set_id, term=$term, def=$definition";
